qemu: split action package into domain-specific files
Split the monolithic qemu.go into block.go (image formats, cache/AIO
modes, export types), cpu.go (CPU features), device.go (char devices,
display, VGA, watchdog), machine.go (accelerators, boot drives, debug
items), network.go (netdev types), and keymap.go (XKB keymap actions).
