+++ pkg/actions/tools/qemu/cpu.go
@@ -1,0 +1,323 @@
+package qemu
+
+import "github.com/carapace-sh/carapace"
+
+// ActionCpuFeatures completes QEMU CPUID feature flags
+//
+//	3dnow (3DNow!)
+//	avx (Advanced Vector Extensions)
+func ActionCpuFeatures() carapace.Action {
+	return carapace.ActionValues(
+		"3dnow",
+		"3dnowext",
+		"3dnowprefetch",
+		"abm",
+		"ace2",
+		"ace2-en",
+		"acpi",
+		"adx",
+		"aes",
+		"amd-no-ssb",
+		"amd-psfd",
+		"amd-ssbd",
+		"amd-stibp",
+		"amx-avx512",
+		"amx-bf16",
+		"amx-bf16-alias",
+		"amx-complex",
+		"amx-complex-alias",
+		"amx-fp16",
+		"amx-fp16-alias",
+		"amx-fp8",
+		"amx-int8",
+		"amx-int8-alias",
+		"amx-movrs",
+		"amx-tf32",
+		"amx-tile",
+		"apic",
+		"apx-nci-ndd-nf",
+		"apxf",
+		"arat",
+		"arch-capabilities",
+		"arch-lbr",
+		"auto-ibrs",
+		"avic",
+		"avx",
+		"avx-ifma",
+		"avx-ne-convert",
+		"avx-vnni",
+		"avx-vnni-int16",
+		"avx-vnni-int8",
+		"avx10",
+		"avx10-128",
+		"avx10-256",
+		"avx10-512",
+		"avx10-vnni-int",
+		"avx2",
+		"avx512-4fmaps",
+		"avx512-4vnniw",
+		"avx512-bf16",
+		"avx512-fp16",
+		"avx512-vp2intersect",
+		"avx512-vpopcntdq",
+		"avx512bitalg",
+		"avx512bw",
+		"avx512cd",
+		"avx512dq",
+		"avx512er",
+		"avx512f",
+		"avx512ifma",
+		"avx512pf",
+		"avx512vbmi",
+		"avx512vbmi2",
+		"avx512vl",
+		"avx512vnni",
+		"bhi-ctrl",
+		"bhi-no",
+		"bmi1",
+		"bmi2",
+		"bus-lock-detect",
+		"cet-ibt",
+		"cet-ss",
+		"cid",
+		"cldemote",
+		"clflush",
+		"clflushopt",
+		"clwb",
+		"clzero",
+		"cmov",
+		"cmp-legacy",
+		"cmpccxadd",
+		"core-capability",
+		"cr8legacy",
+		"cx16",
+		"cx8",
+		"dca",
+		"ddpd-u",
+		"de",
+		"decodeassists",
+		"ds",
+		"ds-cpl",
+		"dtes64",
+		"eraps",
+		"erms",
+		"est",
+		"extapic",
+		"f16c",
+		"fb-clear",
+		"fbsdp-no",
+		"fdp-excptn-only",
+		"flush-l1d",
+		"flushbyasid",
+		"fma",
+		"fma4",
+		"fpu",
+		"fred",
+		"fs-gs-base-ns",
+		"fsgsbase",
+		"fsrc",
+		"fsrm",
+		"fsrs",
+		"full-width-write",
+		"fxsr",
+		"fxsr-opt",
+		"fzrm",
+		"gds-no",
+		"gfni",
+		"gmet",
+		"hle",
+		"ht",
+		"hypervisor",
+		"ia64",
+		"ibpb",
+		"ibpb-brtype",
+		"ibrs",
+		"ibrs-all",
+		"ibs",
+		"intel-psfd",
+		"intel-pt",
+		"intel-pt-lip",
+		"invpcid",
+		"invtsc",
+		"ipred-ctrl",
+		"its-no",
+		"kvm-asyncpf",
+		"kvm-asyncpf-int",
+		"kvm-asyncpf-vmexit",
+		"kvm-hint-dedicated",
+		"kvm-mmu",
+		"kvm-msi-ext-dest-id",
+		"kvm-nopiodelay",
+		"kvm-poll-control",
+		"kvm-pv-eoi",
+		"kvm-pv-ipi",
+		"kvm-pv-sched-yield",
+		"kvm-pv-tlb-flush",
+		"kvm-pv-unhalt",
+		"kvm-steal-time",
+		"kvmclock",
+		"kvmclock-stable-bit",
+		"la57",
+		"lahf-lm",
+		"lam",
+		"lbrv",
+		"lfence-always-serializing",
+		"lkgs",
+		"lm",
+		"lwp",
+		"mca",
+		"mcdt-no",
+		"mce",
+		"md-clear",
+		"mds-no",
+		"misalignsse",
+		"mmx",
+		"mmxext",
+		"monitor",
+		"movbe",
+		"movdir64b",
+		"movdiri",
+		"movrs",
+		"mpx",
+		"msr",
+		"msr-imm",
+		"mtrr",
+		"no-nested-data-bp",
+		"nodeid-msr",
+		"npt",
+		"nrip-save",
+		"null-sel-clr-base",
+		"nx",
+		"osvw",
+		"overflow-recov",
+		"pae",
+		"pat",
+		"pause-filter",
+		"pbe",
+		"pbrsb-no",
+		"pcid",
+		"pclmulqdq",
+		"pcommit",
+		"pdcm",
+		"pdpe1gb",
+		"perfctr-core",
+		"perfctr-nb",
+		"perfmon-v2",
+		"pfthreshold",
+		"pge",
+		"phe",
+		"phe-en",
+		"pks",
+		"pku",
+		"pmm",
+		"pmm-en",
+		"pn",
+		"pni",
+		"popcnt",
+		"prefetchi",
+		"prefetchiti",
+		"pschange-mc-no",
+		"psdp-no",
+		"pse",
+		"pse36",
+		"rdctl-no",
+		"rdpid",
+		"rdrand",
+		"rdseed",
+		"rdtscp",
+		"rfds-clear",
+		"rfds-no",
+		"rrsba-ctrl",
+		"rsba",
+		"rtm",
+		"sbdr-ssdp-no",
+		"sbpb",
+		"sep",
+		"serialize",
+		"sgx",
+		"sgx-aex-notify",
+		"sgx-debug",
+		"sgx-edeccssa",
+		"sgx-exinfo",
+		"sgx-kss",
+		"sgx-mode64",
+		"sgx-provisionkey",
+		"sgx-tokenkey",
+		"sgx1",
+		"sgx2",
+		"sgxlc",
+		"sha-ni",
+		"sha512",
+		"skinit",
+		"skip-l1dfl-vmentry",
+		"sm3",
+		"sm4",
+		"smap",
+		"smep",
+		"smx",
+		"spec-ctrl",
+		"split-lock-detect",
+		"srso-no",
+		"srso-user-kernel-no",
+		"ss",
+		"ssb-no",
+		"ssbd",
+		"sse",
+		"sse2",
+		"sse4.1",
+		"sse4.2",
+		"sse4a",
+		"ssse3",
+		"stibp",
+		"stibp-always-on",
+		"succor",
+		"svm",
+		"svm-lock",
+		"svme-addr-chk",
+		"syscall",
+		"taa-no",
+		"tbm",
+		"tce",
+		"tm",
+		"tm2",
+		"topoext",
+		"tsa-l1-no",
+		"tsa-sq-no",
+		"tsc",
+		"tsc-adjust",
+		"tsc-deadline",
+		"tsc-scale",
+		"tsx-ctrl",
+		"tsx-ldtrk",
+		"umip",
+		"v-vmsave-vmload",
+		"vaes",
+		"verw-clear",
+		"vgif",
+		"virt-ssbd",
+		"vmcb-clean",
+		"vme",
+		"vmx",
+		"vnmi",
+		"vpclmulqdq",
+		"waitpkg",
+		"wbnoinvd",
+		"wdt",
+		"wrmsrns",
+		"x2apic",
+		"xcrypt",
+		"xcrypt-en",
+		"xfd",
+		"xgetbv1",
+		"xop",
+		"xsave",
+		"xsavec",
+		"xsaveerptr",
+		"xsaveopt",
+		"xsaves",
+		"xstore",
+		"xstore-en",
+		"xtpr",
+		"zero-fcs-fds",
+	).Tag("cpu features").Uid("qemu", "cpu-features")
+}
